+++ apps/community/Orders/Loader.php
@@ -14,15 +14,17 @@ class Loader extends \HubletoMain\Core\App
     ]);
   }
 
-  public function installTables(): void
+  public function installTables(int $round): void
   {
-    $mOrder = new \HubletoApp\Community\Orders\Models\Order($this->main);
-    $mOrderProduct = new \HubletoApp\Community\Orders\Models\OrderProduct($this->main);
-    $mHistory = new \HubletoApp\Community\Orders\Models\History($this->main);
+    if ($round == 1) {
+      $mOrder = new \HubletoApp\Community\Orders\Models\Order($this->main);
+      $mOrderProduct = new \HubletoApp\Community\Orders\Models\OrderProduct($this->main);
+      $mHistory = new \HubletoApp\Community\Orders\Models\History($this->main);
 
-    $mOrder->dropTableIfExists()->install();
-    $mOrderProduct->dropTableIfExists()->install();
-    $mHistory->dropTableIfExists()->install();
+      $mOrder->dropTableIfExists()->install();
+      $mOrderProduct->dropTableIfExists()->install();
+      $mHistory->dropTableIfExists()->install();
+    }
   }
 
   public function installDefaultPermissions(): void
